Implememented the Concept board design view. In the process of making the components and their respective logic.

+++ TFG_Projects_APP_Frontend/Entities/Dtos/Components/ComponentCreate.cs
@@ -6,7 +6,7 @@ internal class ComponentCreate
     public int IdType { get; set; }
     public string Title { get; set; } = string.Empty;
 
-    public int PosX { get; set; }
-    public int PosY { get; set; }
+    public double PosX { get; set; }
+    public double PosY { get; set; }
     public string? Content { get; set; }
 }
