+++ TFG_Projects_APP_Frontend/Entities/Dtos/Components/ComponentUpdate.cs
@@ -6,7 +6,7 @@ internal class ComponentUpdate
     public int? IdType { get; set; }
     public string? Title { get; set; }
     public int? IdParent { get; set; }
-    public int? PosX { get; set; }
-    public int? PosY { get; set; }
+    public double? PosX { get; set; }
+    public double? PosY { get; set; }
     public string? Content { get; set; }
 }
